Fix CharUpperA access violation (seems to be an MSVC bug)
This is a strange bug, since we ensure that a CHAR is passed to this
function, but I'll just assume MSVC is stupid and the compiler is trying
inline this function, and in doing so forgets to shorten the length of the
variable. This has got to be the case, since kernelbase!CharUpperA is
receiving a full 8 byte variable rather than 1 byte. Of those 8 bytes, in
my testing, it seems to be all FF except for the least-significant, so
possibly it's uninitialised memory from the function and not a locale error
at all? Although this only seems to be an issue I realised, and I am the
only user of ExplorerEx who doesn't use Windows in a Latin-alphabet
language.

This was my original theory on the topic:

In some cases, SHStripMneumonic[sic] can seemingly give us a wide char, or
an error code in which the upper byte is non-zero. When passed to
CharUpperA (or in this case, "CharUpperCharA"), this causes an
ACCESS VIOLATION as kernelbase!CharUpperA attempts to parse the passed
value as a pointer rather than a character.
This behaviour of CharUpperA  is documented:
https://learn.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-charuppera

The error in question is *almost* benign, even under affected locales.
That I can tell, the two crash paths, under affected locales, are:
  - Attempting to use keyboard accelerators in the start menu
  - Enabling "display scrolling menus" in the personal start menu settings.

Of these, the latter path is significantly more dangerous, as it is a
setting which, if applied, will cause ExplorerEx to immediately crash every
time it is started. The fix is simple: always ensure that the upper byte is
0. The mask "& 0xff" is enough for this purpose.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>