Move all source code to src/ folder to clean up root folder
This is so that people looking at the repository on GitHub can easily see
the README preview.

If you had unstaged changes prior to this commit, there should not be merge
issues OTHER THAN with the vcxproj. You will probably have to manually
reconcile differences there, because vcxproj is an inefficient format.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>