Fix some tray notifications not responding to mouse events
The issue is that XP Explorer, and by extension ExplorerEx, did not
acknowledge the existence of NOTIFYICON version 4, which is used since
Windows Vista. This meant that tray notification events were often not
even sent for newer applications which use version 4.

I referenced the source code of ep_taskbar for this.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>